pa_backup/extracted/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.104.1
greenlet==3.2.4
gTTS==2.5.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.10.1
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.45
starlette==0.27.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
WTForms==3.1.2
python-multipart==0.0.6
itsdangerous==2.2.0
